Split out AppBar/SliverAppBar material tests
